pyproject.toml
[tool.poetry]
name = "cli-translator"
version = "0.1.0"
description = ""
authors = ["Astashov Artem <asst.art@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
pyyaml = "^5.3"
PyInquirer = "^1.0.3"
click = "^7.1.2"
rich = "10.2.2"

[tool.poetry.dev-dependencies]
nose = "^1.3.7"